requirements.txt
# Django e REST Framework
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Coleta de dados
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
playwright==1.40.0
feedparser==6.0.10

# Processamento de dados
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
textblob==0.17.1
yake==0.4.8

# Análise de sentimento
vaderSentiment==3.3.2

# Banco de dados
psycopg2-binary==2.9.9
redis==5.0.1

# Telegram Bot
python-telegram-bot==20.7

# Utilitários
python-dateutil==2.8.2
pytz==2023.3
lxml==4.9.3
html5lib==1.1

# Desenvolvimento
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Produção
gunicorn==21.2.0
whitenoise==6.6.0 